requirements.txt
fastapi==0.115.11
langchain_core==0.3.43
langchain_openai==0.3.8
pydantic==2.10.6
python-dotenv==1.0.1
